+++ app.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-__version__ = "3.3.3"
+__version__ = "3.3.4"
 
 import logging
 import os